pong-sandbox-2: separate classes into their files; add score card functionality
